Implement missing SSZ reference tests.
