+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/AttesterSlashingSszStaticReferenceTest.java
@@ -25,21 +25,21 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.operations.AttestationData;
+import tech.pegasys.artemis.datastructures.operations.AttesterSlashing;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class attestationData extends TestSuite {
+public class AttesterSlashingSszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
-      name = "{index}. ssz_static/AttestationData deserializedAttestationData={0}, root={1}")
+      name = "{index}. ssz_static/AttesterSlashing deserializedAttesterSlashing={0}, root={1}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticAttestationData(AttestationData deserializedAttestationData, Bytes32 root)
+  void processSSZStaticAttesterSlashing(AttesterSlashing deserializedAttesterSlashing, Bytes32 root)
       throws Exception {
-    assertEquals(deserializedAttestationData.hash_tree_root(), root);
+    assertEquals(deserializedAttesterSlashing.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -55,7 +55,7 @@ public class attestationData extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "AttestationData");
-    return sszStaticSetup(path, configPath, AttestationData.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "AttesterSlashing");
+    return sszStaticSetup(path, configPath, AttesterSlashing.class);
   }
 }
