+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/BeaconBlockBodySszStaticReferenceTest.java
@@ -25,22 +25,22 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.blocks.BeaconBlockHeader;
+import tech.pegasys.artemis.datastructures.blocks.BeaconBlockBody;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class beaconBlockHeader extends TestSuite {
+public class BeaconBlockBodySszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
       name =
-          "{index}. ssz_static/BeaconBlockHeader deserializedBeaconBlockHeader={0}, root={1}, signingRoot={2}")
+          "{index}. ssz_static/BeaconBlockBody deserializedBeaconBlockBody={0}, root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticBeaconBlock(BeaconBlockHeader deserializedBeaconBlockHeader, Bytes32 root)
+  void processSSZStaticBeaconBlock(BeaconBlockBody deserializedBeaconBlockBody, Bytes32 root)
       throws Exception {
-    assertEquals(deserializedBeaconBlockHeader.hash_tree_root(), root);
+    assertEquals(deserializedBeaconBlockBody.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -56,7 +56,7 @@ public class beaconBlockHeader extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "BeaconBlockHeader");
-    return sszStaticSetup(path, configPath, BeaconBlockHeader.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "BeaconBlockBody");
+    return sszStaticSetup(path, configPath, BeaconBlockBody.class);
   }
 }
