+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/BeaconStateSszStaticReferenceTest.java
@@ -25,21 +25,21 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.operations.AttesterSlashing;
+import tech.pegasys.artemis.datastructures.state.BeaconState;
+import tech.pegasys.artemis.datastructures.state.BeaconStateImpl;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class attesterSlashing extends TestSuite {
+public class BeaconStateSszStaticReferenceTest extends TestSuite {
 
-  @ParameterizedTest(
-      name = "{index}. ssz_static/AttesterSlashing deserializedAttesterSlashing={0}, root={1}")
+  @ParameterizedTest(name = "{index}. ssz_static/BeaconState root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticAttesterSlashing(AttesterSlashing deserializedAttesterSlashing, Bytes32 root)
+  void processSSZStaticBeaconBlock(BeaconState deserializedBeaconState, Bytes32 root)
       throws Exception {
-    assertEquals(deserializedAttesterSlashing.hash_tree_root(), root);
+    assertEquals(deserializedBeaconState.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -55,7 +55,7 @@ public class attesterSlashing extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "AttesterSlashing");
-    return sszStaticSetup(path, configPath, AttesterSlashing.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "BeaconState");
+    return sszStaticSetup(path, configPath, BeaconStateImpl.class);
   }
 }
