+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/CheckpointSszStaticReferenceTest.java
@@ -25,22 +25,21 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.blocks.BeaconBlockBody;
+import tech.pegasys.artemis.datastructures.state.Checkpoint;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class beaconBlockBody extends TestSuite {
+public class CheckpointSszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
-      name =
-          "{index}. ssz_static/BeaconBlockBody deserializedBeaconBlockBody={0}, root={1}, signingRoot={2}")
+      name = "{index}. ssz_static/Checkpoint deserializedCheckpoint={0}, root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticBeaconBlock(BeaconBlockBody deserializedBeaconBlockBody, Bytes32 root)
+  void processSSZStaticBeaconBlock(Checkpoint deserializedCheckpoint, Bytes32 root)
       throws Exception {
-    assertEquals(deserializedBeaconBlockBody.hash_tree_root(), root);
+    assertEquals(deserializedCheckpoint.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -56,7 +55,7 @@ public class beaconBlockBody extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "BeaconBlockBody");
-    return sszStaticSetup(path, configPath, BeaconBlockBody.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "Checkpoint");
+    return sszStaticSetup(path, configPath, Checkpoint.class);
   }
 }
