+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/DepositSszStaticReferenceTest.java
@@ -25,20 +25,20 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.state.Fork;
+import tech.pegasys.artemis.datastructures.operations.Deposit;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class fork extends TestSuite {
+public class DepositSszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
-      name = "{index}. ssz_static/Fork deserializedFork={0}, root={1}, signingRoot={2}")
+      name = "{index}. ssz_static/Deposit deserializedDeposit={0}, root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticBeaconBlock(Fork deserializedFork, Bytes32 root) throws Exception {
-    assertEquals(deserializedFork.hash_tree_root(), root);
+  void processSSZStaticBeaconBlock(Deposit deserializedDeposit, Bytes32 root) throws Exception {
+    assertEquals(deserializedDeposit.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -54,7 +54,7 @@ public class fork extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "Fork");
-    return sszStaticSetup(path, configPath, Fork.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "Deposit");
+    return sszStaticSetup(path, configPath, Deposit.class);
   }
 }
