+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/ValidatorSszStaticReferenceTest.java
@@ -25,22 +25,20 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.state.HistoricalBatch;
+import tech.pegasys.artemis.datastructures.state.Validator;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class historicalBatch extends TestSuite {
+public class ValidatorSszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
-      name =
-          "{index}. ssz_static/HistoricalBatch deserializedHistoricalBatch={0}, root={1}, signingRoot={2}")
+      name = "{index}. ssz_static/Validator deserializedValidator={0}, root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticBeaconBlock(HistoricalBatch deserializedHistoricalBatch, Bytes32 root)
-      throws Exception {
-    assertEquals(deserializedHistoricalBatch.hash_tree_root(), root);
+  void processSSZStaticBeaconBlock(Validator deserializedValidator, Bytes32 root) throws Exception {
+    assertEquals(deserializedValidator.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -56,7 +54,7 @@ public class historicalBatch extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "HistoricalBatch");
-    return sszStaticSetup(path, configPath, HistoricalBatch.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "Validator");
+    return sszStaticSetup(path, configPath, Validator.class);
   }
 }
