+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/artemis/reference/phase0/ssz_static/VoluntaryExitSszStaticReferenceTest.java
@@ -25,20 +25,22 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import tech.pegasys.artemis.datastructures.operations.Deposit;
+import tech.pegasys.artemis.datastructures.operations.VoluntaryExit;
 import tech.pegasys.artemis.ethtests.TestSuite;
 
 @ExtendWith(BouncyCastleExtension.class)
-public class deposit extends TestSuite {
+public class VoluntaryExitSszStaticReferenceTest extends TestSuite {
 
   @ParameterizedTest(
-      name = "{index}. ssz_static/Deposit deserializedDeposit={0}, root={1}, signingRoot={2}")
+      name =
+          "{index}. ssz_static/VoluntaryExit deserializedVoluntaryExit={0}, root={1}, signingRoot={2}")
   @MethodSource({
     "processMinimal",
     "processMainnet",
   })
-  void processSSZStaticBeaconBlock(Deposit deserializedDeposit, Bytes32 root) throws Exception {
-    assertEquals(deserializedDeposit.hash_tree_root(), root);
+  void processSSZStaticVoluntaryExit(VoluntaryExit deserializedVoluntaryExit, Bytes32 root)
+      throws Exception {
+    assertEquals(deserializedVoluntaryExit.hash_tree_root(), root);
   }
 
   @MustBeClosed
@@ -54,7 +56,7 @@ public class deposit extends TestSuite {
   @MustBeClosed
   static Stream<Arguments> process(String config) throws Exception {
     Path configPath = Paths.get(config);
-    Path path = Paths.get(config, "phase0", "ssz_static", "Deposit");
-    return sszStaticSetup(path, configPath, Deposit.class);
+    Path path = Paths.get(config, "phase0", "ssz_static", "VoluntaryExit");
+    return sszStaticSetup(path, configPath, VoluntaryExit.class);
   }
 }
